created VT100 library from the code supplied from JDN
